backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.10.6
pydantic-settings==2.6.1
python-multipart==0.0.6

# LLM Providers
openai==1.10.0
google-generativeai==0.3.2
langchain==0.1.4

# Telegram
python-telegram-bot==20.8

# Database & ORM
alembic==1.13.1
aiosqlite==0.19.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
